+++ packages/components/src/presentation-components/instance-filter-builder/PresentationInstanceFilterProperty.tsx
@@ -33,7 +33,7 @@ export function PresentationInstanceFilterProperty(props: PresentationInstanceFi
   const { propertyDescription, categoryLabel, fullClassName } = props;
   return (
     <div className="property-item-line">
-      <Tooltip content={propertyDescription.displayLabel} placement="bottom">
+      <Tooltip content={propertyDescription.displayLabel} placement="bottom" zIndex={999999}>
         <div className="property-display-label" title={propertyDescription.displayLabel}>
           {propertyDescription.displayLabel}
         </div>
@@ -44,6 +44,7 @@ export function PresentationInstanceFilterProperty(props: PresentationInstanceFi
             content={<CategoryTooltipContent categoryLabel={categoryLabel} fullClassName={fullClassName} />}
             placement="bottom"
             style={{ textAlign: "left" }}
+            zIndex={999999}
           >
             <div className="badge">
               <Badge className="property-category-badge" backgroundColor={"montecarlo"}>